@@ -23,25 +23,26 @@ environment:
 dependencies:
   cached_network_image: ^3.1.0
   carousel_slider: ^4.0.0
-  cloud_firestore: ^2.5.0
+  cloud_firestore: ^4.9.3
   cupertino_icons: ^1.0.2
   date_format: ^2.0.4
-  firebase_auth: ^3.0.2
-  firebase_core: ^1.5.0
-  firebase_messaging: ^10.0.6
-  firebase_storage: ^10.0.2
+  firebase_auth: ^4.10.1
+  firebase_core: ^2.17.0
+  firebase_storage: ^11.2.8
   flutter:
     sdk: flutter
-  flutter_native_splash: ^1.2.1
-  font_awesome_flutter: ^9.1.0
-  google_fonts: ^2.1.0
-  google_sign_in: ^5.0.7
-  image_picker: ^0.8.3+2
-  intl: ^0.17.0
-  permission_handler: ^8.1.4+2
+  flutter_native_splash: ^2.3.3
+  font_awesome_flutter: ^10.5.0
+  google_fonts: ^6.1.0
+  google_sign_in: ^6.1.5
+  image_picker: ^1.0.4
+  intl: ^0.18.1
+  permission_handler: ^11.0.1
   provider: ^6.0.0
-  syncfusion_flutter_pdfviewer: ^19.1.69-beta
-  velocity_x: ^3.3.0
+  velocity_x: ^4.1.1
+  firebase_core_platform_interface: ^4.8.0
+  firebase_messaging: ^14.6.9
+  syncfusion_flutter_pdfviewer: ^23.1.39
 
 dev_dependencies:
   flutter_test:
